WBT: Move blazor specific code from BuildTestBase into
.. BlazorWasmProjectProvider, and BlazorWasmTestBase. This commit avoids
doing any cleanup, and just moves code around.
