+++ src/mono/wasm/Wasm.Build.Tests/NativeRebuildTests/OptimizationFlagChangeTests.cs
@@ -36,7 +36,7 @@ public class OptimizationFlagChangeTests : NativeRebuildTestsBase
         (buildArgs, BuildPaths paths) = FirstNativeBuild(s_mainReturns42, nativeRelink: true, invariant: false, buildArgs, id);
 
         string mainAssembly = $"{buildArgs.ProjectName}.dll";
-        var pathsDict = GetFilesTable(buildArgs, paths, unchanged: false);
+        var pathsDict = _provider.GetFilesTable(buildArgs, paths, unchanged: false);
         pathsDict.UpdateTo(unchanged: true, mainAssembly, "icall-table.h", "pinvoke-table.h", "driver-gen.c");
         if (cflags.Length == 0)
             pathsDict.UpdateTo(unchanged: true, "pinvoke.o", "corebindings.o", "driver.o");
@@ -62,7 +62,7 @@ public class OptimizationFlagChangeTests : NativeRebuildTestsBase
 
         string output = Rebuild(nativeRelink: true, invariant: false, buildArgs, id, extraBuildArgs: $" {cflags} {ldflags}", verbosity: "normal");
         var newStat = StatFiles(pathsDict.Select(kvp => kvp.Value.fullPath));
-        CompareStat(originalStat, newStat, pathsDict.Values);
+        _provider.CompareStat(originalStat, newStat, pathsDict.Values);
 
         string runOutput = RunAndTestWasmApp(buildArgs, buildDir: _projectDir, expectedExitCode: 42, host: host, id: id);
         TestUtils.AssertSubstring($"Found statically linked AOT module '{Path.GetFileNameWithoutExtension(mainAssembly)}'", runOutput,
