+++ src/mono/wasm/Wasm.Build.Tests/NativeRebuildTests/ReferenceNewAssemblyRebuildTest.cs
@@ -26,7 +26,7 @@ namespace Wasm.Build.NativeRebuild.Tests
             buildArgs = buildArgs with { ProjectName = $"rebuild_tasks_{buildArgs.Config}" };
             (buildArgs, BuildPaths paths) = FirstNativeBuild(s_mainReturns42, nativeRelink, invariant: invariant, buildArgs, id);
 
-            var pathsDict = GetFilesTable(buildArgs, paths, unchanged: false);
+            var pathsDict = _provider.GetFilesTable(buildArgs, paths, unchanged: false);
             pathsDict.UpdateTo(unchanged: true, "corebindings.o");
             if (!buildArgs.AOT) // relinking
                 pathsDict.UpdateTo(unchanged: true, "driver-gen.c");
@@ -52,7 +52,7 @@ namespace Wasm.Build.NativeRebuild.Tests
             Rebuild(nativeRelink, invariant, buildArgs, id);
             var newStat = StatFiles(pathsDict.Select(kvp => kvp.Value.fullPath));
 
-            CompareStat(originalStat, newStat, pathsDict.Values);
+            _provider.CompareStat(originalStat, newStat, pathsDict.Values);
             RunAndTestWasmApp(buildArgs, buildDir: _projectDir, expectedExitCode: 42, host: host, id: id);
         }
     }
